+++ src/lib.rs
@@ -4,3 +4,4 @@ pub mod driver;
 pub mod types;
 pub mod registers;
 pub mod calculate;
+mod validate;