+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - `DEFAULT_FREQUENCY_HZ` constant
+- `assert!` from driver
 
 ## [0.1.0] - 2026-04-22